# api/openapi.yaml
openapi: 3.0.3
info:
  title: Burn Budgeter API
  description: |
    Financial observability API for developers and startups. 
    Track cloud and AI infrastructure burn rates in real-time.
    This API is public and does not require authentication.
  version: 1.0.0
servers:
  - url: http://localhost:8080/v1
    description: Local development server

paths:
  /health:
    get:
      summary: API Health Check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  database: { type: string }
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  database: { type: string }

  /projects:
    post:
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, cash_on_hand]
              properties:
                name: { type: string }
                cash_on_hand: { type: number }
                currency: { type: string, default: "USD" }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'

  /projects/{id}:
    get:
      summary: Get project details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                cash_on_hand: { type: number }
                currency: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  /projects/{id}/stack:
    post:
      summary: Add/Update service in stack
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id, quantity]
              properties:
                service_id: { type: integer }
                quantity: { type: number }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  /projects/{id}/stack/{sid}:
    delete:
      summary: Remove service from stack
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sid
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  /projects/{id}/analyze:
    post:
      summary: Analyze ARCHITECTURE.md
      description: Resets and updates stack based on uploaded file.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                architecture: { type: string, format: binary }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  updated_services:
                    type: array
                    items: { $ref: '#/components/schemas/ProjectService' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  /projects/{id}/export-architecture:
    get:
      summary: Export ARCHITECTURE.md
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            text/markdown:
              schema: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'

  /services:
    get:
      summary: List all services
      parameters:
        - name: provider
          in: query
          description: Optional filter to limit results by provider (e.g. AWS, GCP, OpenAI)
          required: false
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Service' }
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create custom service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, name, unit, price_per_unit]
              properties:
                provider: { type: string }
                name: { type: string }
                unit: { type: string }
                price_per_unit: { type: number }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'

  /services/{id}:
    patch:
      summary: Update service
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                provider: { type: string }
                name: { type: string }
                unit: { type: string }
                price_per_unit: { type: number }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete service
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'

components:
  schemas:
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        cash_on_hand: { type: number }
        currency: { type: string }
        created_at: { type: string, format: date-time }
        burn_rate: { type: number }
        runway: { type: number }
    Service:
      type: object
      properties:
        id: { type: integer }
        provider: { type: string }
        name: { type: string }
        unit: { type: string }
        price_per_unit: { type: number }
    ProjectService:
      type: object
      properties:
        project_id: { type: string, format: uuid }
        service_id: { type: integer }
        quantity: { type: number }
        is_optimized: { type: boolean }
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }

  responses:
    BadRequest:
      description: Bad Request - Invalid JSON or missing parameters
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Unprocessable Entity - Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict - Resource is in use or state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }